various

+++ scripts/generate_h5.py
@@ -10,6 +10,7 @@ parser.add_argument("task_file")
 parser.add_argument("--no_prompt", action="store_true")
 parser.add_argument("--cloud_proc_func", default="extract_red")
 parser.add_argument("--cloud_proc_mod", default="rapprentice.cloud_proc_funcs")
+parser.add_argument("--clouds_only", action="store_true")
 args = parser.parse_args()
 
 
@@ -27,36 +28,34 @@ cloud_proc_mod = importlib.import_module(args.cloud_proc_mod)
 cloud_proc_func = getattr(cloud_proc_mod, args.cloud_proc_func)
 
 
-with open(args.task_file, "r") as fh: task_info = yaml.load(fh)
-
-
-task_name = task_info["name"].strip()
-
 task_dir = osp.dirname(args.task_file)
 
-
+with open(args.task_file, "r") as fh: task_info = yaml.load(fh)
 h5path = osp.join(task_dir, task_info["h5path"].strip())
-if osp.exists(h5path):    
-    os.unlink(h5path)
-        
-    
-hdf = h5py.File(h5path)
+
+
+if args.clouds_only:
+    hdf = h5py.File(h5path, "r+")
+else:
+    if osp.exists(h5path):    
+        os.unlink(h5path)
+    hdf = h5py.File(h5path)
 
     
-bag_infos = task_info["bags"]
+    bag_infos = task_info["bags"]
 
-for (i_bag, bag_info) in enumerate(bag_infos):
-    bag_file = osp.join(task_dir, bag_info["bag_file"])
-    ann_file = osp.join(task_dir, bag_info["annotation_file"])
-    video_dir = bag_info["video_dir"]
+    for (i_bag, bag_info) in enumerate(bag_infos):
+        bag_file = osp.join(task_dir, bag_info["bag_file"])
+        ann_file = osp.join(task_dir, bag_info["annotation_file"])
+        video_dir = bag_info["video_dir"]
 
-    demo_name = bag_info["demo_name"] if "demo_name" in bag_info else "demo%i"%i_bag
+        demo_name = bag_info["demo_name"] if "demo_name" in bag_info else "demo%i"%i_bag
     
-    bag = rosbag.Bag(bag_file)
-    with open(ann_file, "r") as fh: annotations = yaml.load(fh)
+        bag = rosbag.Bag(bag_file)
+        with open(ann_file, "r") as fh: annotations = yaml.load(fh)
     
-    bag_proc.add_bag_to_hdf(bag, annotations, hdf, demo_name)
-    bag_proc.add_rgbd_to_hdf(osp.join(task_dir, video_dir), annotations, hdf, demo_name)
+        bag_proc.add_bag_to_hdf(bag, annotations, hdf, demo_name)
+        bag_proc.add_rgbd_to_hdf(osp.join(task_dir, video_dir), annotations, hdf, demo_name)
